Rename `api` to `apiFunctions`
